The faster simplex from a project in Passau. fsimplex.red is a submodule containing an improved copy of simplex.red. The improvements are mostly improvements of coding in contrast to mathematical improvements.
There is a switch fastsimplex (on by default), which triggers the use of this new code.
"off fastsimplex" will use exactly the old simplex code.

In contrast to the old simplex, we do not use floats but exact arithmetic. Anyway, we are MUCH faster. For instance, the computation time of the netlib afiro example reduces 
from 70 seconds to 1.7 seconds on my MacBook.